+++ internal/cli/cmd/app/app.go
@@ -21,17 +21,17 @@ import (
 	"context"
 	"flag"
 	"fmt"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/gocmd"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/sbom"
+	modConv "github.com/CycloneDX/cyclonedx-gomod/pkg/sbom/convert/module"
+	pkgConv "github.com/CycloneDX/cyclonedx-gomod/pkg/sbom/convert/pkg"
 	"os"
 	"path/filepath"
 	"strings"
 
 	cdx "github.com/CycloneDX/cyclonedx-go"
 	cliUtil "github.com/CycloneDX/cyclonedx-gomod/internal/cli/util"
-	"github.com/CycloneDX/cyclonedx-gomod/internal/gocmd"
-	"github.com/CycloneDX/cyclonedx-gomod/internal/gomod"
-	"github.com/CycloneDX/cyclonedx-gomod/internal/sbom"
-	modConv "github.com/CycloneDX/cyclonedx-gomod/internal/sbom/convert/module"
-	pkgConv "github.com/CycloneDX/cyclonedx-gomod/internal/sbom/convert/pkg"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/gomod"
 	"github.com/peterbourgon/ff/v3/ffcli"
 	"github.com/rs/zerolog/log"
 )
