+++ internal/cli/cmd/bin/bin.go
@@ -21,6 +21,8 @@ import (
 	"context"
 	"flag"
 	"fmt"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/sbom"
+	modConv "github.com/CycloneDX/cyclonedx-gomod/pkg/sbom/convert/module"
 	"path/filepath"
 	"strings"
 	"time"
@@ -29,9 +31,7 @@ import (
 
 	cdx "github.com/CycloneDX/cyclonedx-go"
 	cliUtil "github.com/CycloneDX/cyclonedx-gomod/internal/cli/util"
-	"github.com/CycloneDX/cyclonedx-gomod/internal/gomod"
-	"github.com/CycloneDX/cyclonedx-gomod/internal/sbom"
-	modConv "github.com/CycloneDX/cyclonedx-gomod/internal/sbom/convert/module"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/gomod"
 	"github.com/peterbourgon/ff/v3/ffcli"
 	"github.com/rs/zerolog/log"
 )
