+++ internal/cli/cmd/mod/mod.go
@@ -22,14 +22,14 @@ import (
 	"errors"
 	"flag"
 	"fmt"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/gocmd"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/sbom"
+	modConv "github.com/CycloneDX/cyclonedx-gomod/pkg/sbom/convert/module"
 	"io"
 
 	cdx "github.com/CycloneDX/cyclonedx-go"
 	cliUtil "github.com/CycloneDX/cyclonedx-gomod/internal/cli/util"
-	"github.com/CycloneDX/cyclonedx-gomod/internal/gocmd"
-	"github.com/CycloneDX/cyclonedx-gomod/internal/gomod"
-	"github.com/CycloneDX/cyclonedx-gomod/internal/sbom"
-	modConv "github.com/CycloneDX/cyclonedx-gomod/internal/sbom/convert/module"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/gomod"
 	"github.com/peterbourgon/ff/v3/ffcli"
 	"github.com/rs/zerolog/log"
 )
