+++ pkg/gomod/binary_test.go
