+++ pkg/gomod/module.go
@@ -22,13 +22,13 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/gocmd"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/util"
 	"io"
 	"path/filepath"
 	"sort"
 	"strings"
 
-	"github.com/CycloneDX/cyclonedx-gomod/internal/gocmd"
-	"github.com/CycloneDX/cyclonedx-gomod/internal/util"
 	"github.com/rs/zerolog/log"
 	"golang.org/x/mod/semver"
 	"golang.org/x/mod/sumdb/dirhash"
