+++ pkg/gomod/stdlib.go
@@ -19,9 +19,8 @@ package gomod
 
 import (
 	"fmt"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/gocmd"
 	"path/filepath"
-
-	"github.com/CycloneDX/cyclonedx-gomod/internal/gocmd"
 )
 
 // StdlibModulePath defines the path used for Go's standard library module.
