+++ pkg/gomod/vendor.go
@@ -22,13 +22,12 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/gocmd"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/util"
 	"github.com/rs/zerolog/log"
 	"io"
 	"path/filepath"
 	"strings"
-
-	"github.com/CycloneDX/cyclonedx-gomod/internal/gocmd"
-	"github.com/CycloneDX/cyclonedx-gomod/internal/util"
 )
 
 // IsVendoring determines whether of not the module at moduleDir is vendoring its dependencies.
