+++ pkg/gomod/version_test.go
