+++ pkg/license/license_test.go
@@ -20,7 +20,7 @@ package license
 import (
 	"testing"
 
-	"github.com/CycloneDX/cyclonedx-gomod/internal/gomod"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/gomod"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
