+++ pkg/sbom/convert/module/module.go
@@ -21,13 +21,13 @@ import (
 	"encoding/base64"
 	"errors"
 	"fmt"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/license"
+	pkgConv "github.com/CycloneDX/cyclonedx-gomod/pkg/sbom/convert/pkg"
 	"regexp"
 	"strings"
 
 	cdx "github.com/CycloneDX/cyclonedx-go"
-	"github.com/CycloneDX/cyclonedx-gomod/internal/gomod"
-	"github.com/CycloneDX/cyclonedx-gomod/internal/license"
-	pkgConv "github.com/CycloneDX/cyclonedx-gomod/internal/sbom/convert/pkg"
+	"github.com/CycloneDX/cyclonedx-gomod/pkg/gomod"
 	"github.com/rs/zerolog/log"
 )
