+++ pkg/sbom/sbom_test.go
